Move Topsters 2 import button to imports section
